MAINT: Merge public and private dtype API as much as possible
This merges header definitions that are private (enough) so that
we use the same definitions within NumPy as externally made available
through the experimental dtype API header.

Tested with string and mpfdtype from the experimental dtype API.
